services/recreate-instances/requirements.txt
google-cloud-compute
python-dotenv